+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.2.0 - 2024/11/10
 
+* [Rustdesk Client](https://rustdesk.com/docs/en/client/)
+
 ## v0.1.0 - 2024/11/09
 
 * Debian 12 (Bookworm)